+++ dojo/product/__init__.py
@@ -1,0 +1,1 @@
+import dojo.product.admin  # noqa: F401
